+++ tree-sitter-langs.el
@@ -124,6 +124,7 @@ See `tree-sitter-langs-repos'."
                 (jsonc-mode      . json)
                 (julia-mode      . julia)
                 (ocaml-mode      . ocaml)
+                (perl-mode       . perl)
                 (php-mode        . php)
                 (prisma-mode     . prisma)
                 (python-mode     . python)