+++ queries/go/highlights.scm
@@ -64,6 +64,7 @@
   "|"
   "|="
   "||"
+  "~"
 ] @operator
 
 ; Keywords
@@ -116,6 +117,7 @@
   (true)
   (false)
   (nil)
+  (iota)
 ] @constant.builtin
 
 (comment) @comment