+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Added initial support for Perl
 - Added highlighting for `haskell-mode`.
 
 ## 0.12.2 - 2022-09-11