+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Updated `php` highlighting with assorted fixes.
 
 ## 0.12.3 - 2022-09-15
 - Added initial support for Perl